Work TDS example, fix framework bugs

+++ src/TopDownShooter/Input/ClientInput.cs
@@ -5,7 +5,7 @@ namespace TopDownShooter.Input;
 [MemoryPackable]
 partial class ClientInput
 {
-    public int Horizontal;
-    public int Vertical;
+    public sbyte Horizontal;
+    public sbyte Vertical;
     public bool Start;
 }
